requirements.txt
# Core
boto3>=1.34.0
botocore>=1.34.0

# Streaming & Batch Processing
pyspark==3.5.0
pyarrow>=14.0.0
s3fs>=2024.1.0

# Data Quality
great-expectations>=0.18.0

# Orchestration
apache-airflow==2.8.0
apache-airflow-providers-amazon>=8.15.0

# Data & Utilities
pandas>=2.1.0
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
ruff>=0.2.0
mypy>=1.8.0
sqlfluff>=2.3.0
